third_party/blink/web_tests/platform/generic/fast/layers/negative-scroll-positions-expected.txt

This tests that negative scrollTop and scrollLeft values are clamped to zero.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS attemptScroll(0, 0) is '0,0'
PASS attemptScroll(50, 50) is '50,50'
PASS attemptScroll(-50, -50) is '0,0'
PASS successfullyParsed is true

TEST COMPLETE
